Simplify route logging, fix DateTime format, and extract count/getById modules
- Remove LOGGED_API_PATH_PATTERNS and minimatch — use ROUTE_PATTERNS for both logging eligibility and pattern extraction
- Trim ROUTE_PATTERNS to only non-GET routes, fix ordering (literal before parameterized)
- Fix Tinybird DateTime type mismatch by formatting timestamps as YYYY-MM-DD HH:MM:SS
- Extract getApiLogsCount into get-api-logs-count.ts with proper data extraction from Tinybird response
- Extract getApiLogById into get-api-log.ts
- Use useApiLogsCount hook with getQueryString in logs table and filters
