+++ apps/web/app/api/api-logs/[logId]/route.ts
@@ -1,5 +1,5 @@
 import { enrichApiLogs } from "@/lib/api-logs/enrich-api-logs";
-import { getApiLogById } from "@/lib/api-logs/get-api-logs";
+import { getApiLogById } from "@/lib/api-logs/get-api-log";
 import { DubApiError } from "@/lib/api/errors";
 import { withWorkspace } from "@/lib/auth/workspace";
 import { NextResponse } from "next/server";
