+++ apps/web/app/api/api-logs/count/route.ts
@@ -1,5 +1,5 @@
 import { getApiLogsStartDate } from "@/lib/api-logs/api-log-retention";
-import { getApiLogsCount } from "@/lib/api-logs/get-api-logs";
+import { getApiLogsCount } from "@/lib/api-logs/get-api-logs-count";
 import { getApiLogsCountQuerySchema } from "@/lib/api-logs/schemas";
 import { withWorkspace } from "@/lib/auth/workspace";
 import { NextResponse } from "next/server";
@@ -9,13 +9,13 @@ export const GET = withWorkspace(
   async ({ workspace, searchParams }) => {
     const filters = getApiLogsCountQuerySchema.parse(searchParams);
 
-    const result = await getApiLogsCount({
+    const count = await getApiLogsCount({
       ...filters,
       workspaceId: workspace.id,
       start: getApiLogsStartDate(workspace.plan),
     });
 
-    return NextResponse.json(result);
+    return NextResponse.json(count);
   },
   {
     requiredPermissions: ["workspaces.read"],
