+++ apps/web/lib/api-logs/constants.ts
@@ -1,16 +1,5 @@
-export const LOGGED_API_PATH_PATTERNS = [
-  "/api/track/lead",
-  "/api/track/sale",
-  "/api/track/open",
-  "/api/partners/**",
-  "/api/tokens/embed/referrals",
-  "/api/customers/**",
-  "/api/links/**",
-  "/api/commissions/**",
-  "/api/bounties/**",
-] as const;
-
 // Route patterns for parameterized path matching.
+// Used both for logging eligibility and route pattern extraction.
 // Order matters: more specific patterns must come before less specific ones.
 export const ROUTE_PATTERNS = [
   // Track
@@ -18,64 +7,35 @@ export const ROUTE_PATTERNS = [
   "api/track/sale",
   "api/track/open",
 
-  // Links
-  "api/links/:linkId/dashboard",
-  "api/links/:linkId/transfer",
-  "api/links/:linkId",
-  "api/links/bulk",
-  "api/links/count",
-  "api/links/exists",
-  "api/links/export",
-  "api/links/info",
-  "api/links/upsert",
-  "api/links",
-
   // Partners
-  "api/partners/:partnerId/application-risks",
-  "api/partners/:partnerId/comments/count",
-  "api/partners/:partnerId/comments",
-  "api/partners/:partnerId/cross-program-summary",
-  "api/partners/:partnerId",
-  "api/partners/analytics",
-  "api/partners/ban",
-  "api/partners/count",
-  "api/partners/deactivate",
-  "api/partners/export",
   "api/partners/links/upsert",
   "api/partners/links",
-  "api/partners/platforms/callback",
+  "api/partners/ban",
+  "api/partners/deactivate",
+  "api/partners/:partnerId",
   "api/partners",
 
+  // Links
+  "api/links/bulk",
+  "api/links/upsert",
+  "api/links/:linkId",
+  "api/links",
+
   // Customers
-  "api/customers/:id/activity",
-  "api/customers/:id/stripe-invoices",
   "api/customers/:id",
-  "api/customers/count",
-  "api/customers/export",
-  "api/customers/search-stripe",
   "api/customers",
 
   // Commissions
-  "api/commissions/:commissionId",
   "api/commissions/bulk",
-  "api/commissions/count",
-  "api/commissions/export",
-  "api/commissions/timeseries",
+  "api/commissions/:commissionId",
   "api/commissions",
 
   // Bounties
   "api/bounties/:bountyId/submissions/:submissionId/approve",
   "api/bounties/:bountyId/submissions/:submissionId/reject",
-  "api/bounties/:bountyId/submissions",
-  "api/bounties/:bountyId/sync-social-metrics",
-  "api/bounties/:bountyId",
-  "api/bounties/count/submissions",
-  "api/bounties",
 
   // Tokens
   "api/tokens/embed/referrals",
-  "api/tokens/:id",
-  "api/tokens",
 ] as const;
 
 export const HTTP_STATUS_CODES = [
