+++ apps/web/lib/api-logs/get-api-logs-count.ts
@@ -1,0 +1,48 @@
+import { tb } from "@/lib/tinybird";
+import * as z from "zod/v4";
+import { prefixWorkspaceId } from "../api/workspaces/workspace-id";
+import {
+  apiLogCountFilterSchemaTB,
+  apiLogCountResponseSchemaTB,
+} from "./schemas";
+
+type GetApiLogsCountParams = z.infer<typeof apiLogCountFilterSchemaTB>;
+
+export async function getApiLogsCount(params: GetApiLogsCountParams) {
+  const {
+    workspaceId,
+    routePattern,
+    method,
+    statusCode,
+    tokenId,
+    requestId,
+    start,
+    groupBy,
+  } = params;
+
+  const baseParams = {
+    workspaceId: prefixWorkspaceId(workspaceId),
+    ...(groupBy && { groupBy }),
+    ...(routePattern && { routePattern }),
+    ...(method && { method }),
+    ...(statusCode && { statusCode }),
+    ...(tokenId && { tokenId }),
+    ...(requestId && { requestId }),
+    ...(start && { start }),
+    end: new Date().toISOString().replace("T", " ").slice(0, 19),
+  };
+
+  const pipe = tb.buildPipe({
+    pipe: "count_api_logs",
+    parameters: apiLogCountFilterSchemaTB,
+    data: z.any(),
+  });
+
+  const result = await pipe(baseParams);
+
+  if (groupBy === "routePattern") {
+    return z.array(apiLogCountResponseSchemaTB.routePattern).parse(result.data);
+  }
+
+  return result.data[0]?.count ?? 0;
+}
