+++ packages/tinybird/datasources/dub_api_logs.datasource
@@ -17,5 +17,5 @@ SCHEMA >
 
 ENGINE "MergeTree"
 ENGINE_PARTITION_KEY "toYYYYMM(timestamp)"
-ENGINE_SORTING_KEY "workspace_id, path, timestamp"
+ENGINE_SORTING_KEY "workspace_id, timestamp"
 ENGINE_TTL "toDateTime(timestamp) + toIntervalDay(90)"
